requirements.txt
pyyaml
tqdm
omegaconf
hydra-core
numpy
matplotlib
pillow
prettytable
opencv-python
scikit-learn
